footprinter/requirements.txt
dnspython
whois
requests
sys
argparse
socket
